refactor: some more cleanup and lint fixes
