+++ src/token/cose/encrypt/encrypt/mod.rs
@@ -1,4 +1,5 @@
 use alloc::rc::Rc;
+use alloc::vec::Vec;
 use core::cell::RefCell;
 
 use coset::{CoseEncrypt, CoseEncryptBuilder, EncryptionContext, Header};
